backend/requirements.txt
# LiveTest backend – Python 3.12 compatible
imutils==0.5.4
opencv-python-headless==4.12.0.88
SQLAlchemy==2.0.30
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.11.7
pydantic[email]
passlib==1.7.4
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
psycopg2-binary==2.9.10
numpy==2.1.1
Pillow==10.4.0
Faker==22.5.0
scipy==1.14.1
rich==13.9.2
python-multipart==0.0.9
